examples: add pacmirror variant that uses rsync
The version using urlretrieve is overly slow, so make a version that
popens rsync and sends it a \0-delimtited list of files to sync.  Since
rsync is already intelligent about such things, just send it a list of
all files from the db, and let it figure out if they need to actually be
downloaded or not.